backend_requirements.txt
# Backend Requirements for SportsBuddy API
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
redis==5.0.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
geopy==2.4.0
shapely==2.0.2
pytest==7.4.3
httpx==0.25.1
locust==2.17.0
streamlit==1.32.0
plotly==5.19.0
folium==0.16.0
streamlit-folium==0.19.0
